+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Updated
 
+- [Desktop: Collaborators](https://app.asana.com/1/355454733208534/project/1209462826205010/task/1209998425917886) 
+
 - [Mobile: Home](https://app.asana.com/1/355454733208534/project/1209462826205010/task/1210004012538696) 
 - [Mobile: Jobs](https://app.asana.com/1/355454733208534/project/1209462826205010/task/1210004012538692) 
 - [Mobile: Single Collaborator](https://app.asana.com/1/355454733208534/project/1209462826205010/task/1210004016350637)